metachat-all-services/metachat-mood-analysis-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.1
confluent-kafka==2.3.0
torch==2.1.0
transformers==4.35.2
sentencepiece==0.1.99
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
prometheus-client==0.19.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pymorphy3==2.0.0
nltk==3.8.1
httpx==0.25.2
